BM_LegalCopilot/requirements.txt
streamlit
openai
PyMuPDF
python-dotenv